fix(DemoDataSource): incorrect response for some motor packets
