+++ lib/domain/data_source/models/package/incoming/motor_speed.dart
@@ -2,7 +2,7 @@ import 'package:pixel_app_flutter/domain/data_source/models/package/data_source_
 import 'package:pixel_app_flutter/domain/data_source/models/package/mixins/motor_index_mixin.dart';
 
 class MotorSpeedIncomingDataSourcePackage
-    extends UInt16WithStatusIncomingDataSourcePackage with MotorIndexMixin {
+    extends Uint16WithStatusIncomingDataSourcePackage with MotorIndexMixin {
   MotorSpeedIncomingDataSourcePackage(super.source);
 
   @override
